+++ Panels/index.js
@@ -2,7 +2,7 @@
  * Panels provides a way to manage different screens of an app.
  *
  * @module agate/Panels
- * @exports ActivityPanels
+ * @exports BreadcrumbPanels
  * @exports Breadcrumb
  * @exports Panel
  * @exports Panels
@@ -10,15 +10,15 @@
  * @exports TabbedPanels
  */
 
-import ActivityPanels from './ActivityPanels';
 import Breadcrumb from './Breadcrumb';
+import BreadcrumbPanels from './BreadcrumbPanels';
 import Panel from './Panel';
 import {Panels, PanelsBase} from './Panels';
 import TabbedPanels from './TabbedPanels';
 
 export default Panels;
 export {
-	ActivityPanels,
+	BreadcrumbPanels,
 	Breadcrumb,
 	Panel,
 	Panels,
